chore: init clarity abstracting classes and functions
